Save IP address instead of hostname
One of the Tucson machines being used to store teststand raw files into
the DBB only has name localhost.localdomain.   For this and other
reasons (like known max length), change identification of machine from
where the raw files were saved to DBB from the hostname to the IP
address.

Because of the localhost.localdomain, the code could not use the
normal method for getting the IP address from the hostname.