Fix transfer process not always cleanly ending
Files in delivery area on NCSA gateway machine were there as temp files but
then disappeared instead of being renamed real filename.  Further debugging
discovered that curl subprocess was not itself always ending but the code
moved on and Python garbage collection cleaned it up.   When the tarfile
close occurred, it did not seem to close the subprocess pipe it was using
for output.